# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LineageEdge": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "edge_id": {
            "title": "Edge Id",
            "type": "string"
          },
          "edge_type": {
            "enum": [
              "derived_from",
              "converted_to",
              "candidate_for",
              "mapped_to",
              "reviewed_by",
              "influenced_by",
              "validated_against",
              "rendered_as",
              "contained_in",
              "verified_by"
            ],
            "title": "Edge Type",
            "type": "string"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Ids",
            "type": "array"
          },
          "metadata": {
            "title": "Metadata",
            "type": "object"
          },
          "source_node_id": {
            "title": "Source Node Id",
            "type": "string"
          },
          "target_node_id": {
            "title": "Target Node Id",
            "type": "string"
          }
        },
        "required": [
          "edge_id",
          "source_node_id",
          "target_node_id",
          "edge_type"
        ],
        "title": "LineageEdge",
        "type": "object"
      },
      "LineageEvidence": {
        "additionalProperties": false,
        "properties": {
          "artifact_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Path"
          },
          "block_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Block Id"
          },
          "evidence_id": {
            "title": "Evidence Id",
            "type": "string"
          },
          "evidence_type": {
            "enum": [
              "source_text",
              "source_path",
              "adapter_trace",
              "candidate_value",
              "mapping_evidence",
              "review_note",
              "knowledge_rule",
              "manifest_hash",
              "contract_check"
            ],
            "title": "Evidence Type",
            "type": "string"
          },
          "metadata": {
            "title": "Metadata",
            "type": "object"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha256"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          }
        },
        "required": [
          "evidence_id",
          "evidence_type"
        ],
        "title": "LineageEvidence",
        "type": "object"
      },
      "LineageGraph": {
        "additionalProperties": false,
        "properties": {
          "doc_id": {
            "title": "Doc Id",
            "type": "string"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/LineageEdge"
            },
            "title": "Edges",
            "type": "array"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/LineageEvidence"
            },
            "title": "Evidence",
            "type": "array"
          },
          "generated_at": {
            "title": "Generated At",
            "type": "string"
          },
          "graph_id": {
            "title": "Graph Id",
            "type": "string"
          },
          "lineage_version": {
            "default": "1.0",
            "title": "Lineage Version",
            "type": "string"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/LineageNode"
            },
            "title": "Nodes",
            "type": "array"
          },
          "package_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Package Id"
          },
          "schema_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Id"
          },
          "summary": {
            "title": "Summary",
            "type": "object"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "template_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Id"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "graph_id",
          "doc_id",
          "generated_at",
          "nodes",
          "edges",
          "evidence"
        ],
        "title": "LineageGraph",
        "type": "object"
      },
      "LineageNode": {
        "additionalProperties": false,
        "properties": {
          "artifact_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Path"
          },
          "block_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Block Id"
          },
          "chunk_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chunk Id"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "confidence_tier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Tier"
          },
          "doc_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doc Id"
          },
          "field_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Name"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "metadata": {
            "title": "Metadata",
            "type": "object"
          },
          "node_id": {
            "title": "Node Id",
            "type": "string"
          },
          "node_type": {
            "enum": [
              "external_field",
              "adapter_trace",
              "uir_block",
              "field_candidate",
              "mapping_decision",
              "review_decision",
              "knowledge_pack",
              "schema_field",
              "canonical_field",
              "rendered_artifact",
              "chunk",
              "package_manifest_entry",
              "consumer_contract"
            ],
            "title": "Node Type",
            "type": "string"
          },
          "review_required_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Required Reason"
          },
          "risk_flags": {
            "items": {
              "type": "string"
            },
            "title": "Risk Flags",
            "type": "array"
          },
          "schema_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Id"
          },
          "schema_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Version"
          },
          "status": {
            "default": "informational",
            "enum": [
              "accepted",
              "review_required",
              "blocked",
              "failed",
              "warning",
              "informational"
            ],
            "title": "Status",
            "type": "string"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "template_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Id"
          },
          "template_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Version"
          }
        },
        "required": [
          "node_id",
          "node_type",
          "label"
        ],
        "title": "LineageNode",
        "type": "object"
      },
      "LineageQueryResult": {
        "additionalProperties": false,
        "properties": {
          "direction": {
            "enum": [
              "upstream",
              "downstream",
              "both"
            ],
            "title": "Direction",
            "type": "string"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/LineageEdge"
            },
            "title": "Edges",
            "type": "array"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/LineageEvidence"
            },
            "title": "Evidence",
            "type": "array"
          },
          "max_depth": {
            "title": "Max Depth",
            "type": "integer"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/LineageNode"
            },
            "title": "Nodes",
            "type": "array"
          },
          "root_node_id": {
            "title": "Root Node Id",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "object"
          }
        },
        "required": [
          "root_node_id",
          "direction",
          "max_depth",
          "nodes",
          "edges",
          "evidence"
        ],
        "title": "LineageQueryResult",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "SchemaPack Agent",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/tasks/{task_id}/lineage": {
      "get": {
        "operationId": "get_task_lineage_api_v1_tasks__task_id__lineage_get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineageGraph"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Task Lineage",
        "tags": [
          "lineage"
        ]
      }
    },
    "/api/v1/tasks/{task_id}/lineage/artifacts/{artifact_path}": {
      "get": {
        "operationId": "get_artifact_lineage_api_v1_tasks__task_id__lineage_artifacts__artifact_path__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "artifact_path",
            "required": true,
            "schema": {
              "title": "Artifact Path",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "direction",
            "required": false,
            "schema": {
              "default": "both",
              "enum": [
                "upstream",
                "downstream",
                "both"
              ],
              "title": "Direction",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "max_depth",
            "required": false,
            "schema": {
              "default": 8,
              "maximum": 32,
              "minimum": 1,
              "title": "Max Depth",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineageQueryResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Artifact Lineage",
        "tags": [
          "lineage"
        ]
      }
    },
    "/api/v1/tasks/{task_id}/lineage/chunks/{chunk_id}": {
      "get": {
        "operationId": "get_chunk_lineage_api_v1_tasks__task_id__lineage_chunks__chunk_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "chunk_id",
            "required": true,
            "schema": {
              "title": "Chunk Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "direction",
            "required": false,
            "schema": {
              "default": "upstream",
              "enum": [
                "upstream",
                "downstream",
                "both"
              ],
              "title": "Direction",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "max_depth",
            "required": false,
            "schema": {
              "default": 8,
              "maximum": 32,
              "minimum": 1,
              "title": "Max Depth",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineageQueryResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Chunk Lineage",
        "tags": [
          "lineage"
        ]
      }
    },
    "/api/v1/tasks/{task_id}/lineage/fields/{field_name}": {
      "get": {
        "operationId": "get_field_lineage_api_v1_tasks__task_id__lineage_fields__field_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "field_name",
            "required": true,
            "schema": {
              "title": "Field Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "direction",
            "required": false,
            "schema": {
              "default": "upstream",
              "enum": [
                "upstream",
                "downstream",
                "both"
              ],
              "title": "Direction",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "max_depth",
            "required": false,
            "schema": {
              "default": 8,
              "maximum": 32,
              "minimum": 1,
              "title": "Max Depth",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineageQueryResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Field Lineage",
        "tags": [
          "lineage"
        ]
      }
    },
    "/api/v1/tasks/{task_id}/lineage/summary": {
      "get": {
        "operationId": "get_lineage_summary_api_v1_tasks__task_id__lineage_summary_get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Lineage Summary Api V1 Tasks  Task Id  Lineage Summary Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Lineage Summary",
        "tags": [
          "lineage"
        ]
      }
    },
...
  }
}